Refactor ResourceMonitor into monitor.js
ResourceMonitor has been moved from app.js into monitor.js and
renamed to simply Monitor. It also gains the ability to fine-
grained control the log level, although in a pretty crude way.

Monitor.debug - bugs in PS's engine that we need to fix
someday - by default, these aren't reported

Monitor.warn - issues that suggest an attacker or poorly
coded custom client - these should probably be looked at

Monitor.notice - issues that staff probably want to know.
These are usually also written to the Staff room.
