+++ gulpfile.js
@@ -15,7 +15,7 @@ var fileCache = new CacheSwap({tmpDir: '', cacheDirName: 'gulp-cache'});
 
 var globals = {};
 var globalList = [
-	'Config', 'ResourceMonitor', 'toId', 'Tools', 'LoginServer', 'Users', 'Rooms', 'Verifier',
+	'Config', 'Monitor', 'toId', 'Tools', 'LoginServer', 'Users', 'Rooms', 'Verifier',
 	'CommandParser', 'Simulator', 'Tournaments', 'Dnsbl', 'Cidr', 'Sockets', 'TeamValidator',
 	'Ladders'
 ];
